ml-model/mongo/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.5.0
certifi>=2023.0.0
python-dotenv==1.0.0
PyJWT==2.8.0
bcrypt==4.1.2
authlib==1.3.0
requests==2.31.0
python-multipart==0.0.6
razorpay==1.4.1
